refactor: 스타일 파일 분리, 파일 위치 변경
