+++ packages/client-todo-list/src/components/Todo/TodoList/TodoList.tsx
@@ -1,10 +1,11 @@
 /* eslint-disable react/no-array-index-key */
 import { FC, useCallback } from 'react';
 
-import { css } from '@emotion/react';
 import classNames from 'classnames';
 
 import { useTodoStore } from 'hooks/useTodoStore';
+
+import * as S from './style';
 
 interface Props {}
 const TodoList: FC<Props> = () => {
@@ -15,42 +16,7 @@ const TodoList: FC<Props> = () => {
   }, [folders, selectedFolderIndex]);
 
   return (
-    <div
-      css={(theme) => css`
-        display: flex;
-        margin: 20px;
-        > div {
-          display: flex;
-          flex-direction: column;
-          gap: 10px;
-        }
-        div:nth-of-type(1) {
-          flex: 1;
-        }
-        div:nth-of-type(2) {
-          flex: 5;
-        }
-        h1 {
-          ${theme.typo.titleL}
-        }
-        ul {
-          display: flex;
-          flex-direction: column;
-          gap: 5px;
-          li {
-            display: flex;
-            gap: 10px;
-            list-style: inside;
-            &.selected {
-              color: ${theme.color.blue500};
-            }
-            .btn-remove {
-              color: ${theme.color.red500};
-            }
-          }
-        }
-      `}
-    >
+    <div css={S.wrapperStyle}>
       <div>
         <h1>Folder</h1>
         <ul>
