+++ packages/client-todo-list/src/components/Todo/TodoList/index.ts
@@ -1,0 +1,1 @@
+export { default as TodoList } from './TodoList';
