+++ packages/client-todo-list/src/components/Todo/TodoListInputs/index.ts
@@ -1,0 +1,1 @@
+export { default as TodoListInputs } from './TodoListInputs';
