+++ packages/client-todo-list/src/components/Todo/index.ts
@@ -1,0 +1,2 @@
+export * from './TodoList';
+export * from './TodoListInputs';
