+++ packages/client-todo-list/src/components/common/Input/Input.tsx
@@ -1,0 +1,18 @@
+import { FC, InputHTMLAttributes, useId } from 'react';
+
+import * as S from './style';
+
+interface Props extends InputHTMLAttributes<HTMLInputElement> {
+  label?: string;
+}
+const Input: FC<Props> = ({ label, ...otherProps }) => {
+  const id = useId();
+  return (
+    <div css={S.wrapperStyle}>
+      {label && <label htmlFor={id}>folder</label>}
+      <input id={id} {...otherProps} />
+    </div>
+  );
+};
+
+export default Input;
