--- packages/client-todo-list/src/components/index.ts
@@ -1,3 +1,0 @@
-export { default as TodoList } from './TodoList';
-export { default as TodoListInputs } from './TodoListInputs';
-export { default as Input } from './Input';
