+++ packages/client-todo-list/src/pages/useSyncExternalStore/index.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react';
 
 import { css } from '@emotion/react';
 
-import { TodoListInputs, TodoList } from 'components';
+import { TodoList, TodoListInputs } from 'components/Todo';
 
 interface Props {}
 const UseSyncExternalStorePage: FC<Props> = () => {
